bench: move all monitors to namepspace
