+++ BlasII.Randomizer.Benchmarks/Monitors/AverageSuccessTimeMonitor.cs
@@ -1,0 +1,25 @@
+﻿
+namespace BlasII.Randomizer.Benchmarks.Monitors;
+
+public class AverageSuccessTimeMonitor : BaseMonitor
+{
+    public override string DisplayName { get; } = "Avg. Time (Successful)";
+
+    protected override void HandleResult(MonitorStatus status, TimeSpan time, bool result)
+    {
+        if (result)
+        {
+            status.SuccessfulTime += time;
+            status.SuccessfulAttempts++;
+        }
+    }
+
+    protected override string FormatResult(MonitorStatus status)
+    {
+        if (status.SuccessfulAttempts == 0)
+            throw new DivideByZeroException("Attempts was zero");
+
+        double time = status.SuccessfulTime.TotalMilliseconds / status.SuccessfulAttempts;
+        return $"{time:F2} ms";
+    }
+}
